Class

+++ Class/Class.Binary/CountWriteOperate.cs
@@ -4,9 +4,9 @@ public class CountWriteOperate : WriteOperate
 {
     public virtual Write Write { get; set; }
 
-    public override bool ExecuteByte(int value)
+    public override bool ExecuteByte(long value)
     {
-        int index;
+        long index;
         index = this.Write.Index;
         index = index + 1;
         this.Write.Index = index;
